added json encoding to link and KC; moved encoder to own module
